pyproject.toml
[project]
name = "exparso"
version = "0.0.3"
description = "Analyzing and parsing documents"
requires-python = ">=3.10"
license = "MIT"
keywords = [ "langchain", "openai", "pdfplumber", "pillow", "openpyxl", "pdf"]
dependencies = [
    "langchain>=0.3.14",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "pydantic==2.9.2",
    "openpyxl>=3.1.5",
    "pdfplumber>=0.11.4",
    "pillow>=10.4.0",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
    "aiofiles>=23.2.1",
    "python-dotenv>=1.0.0",
    "unstructured[pdf]==0.18.15",
    "pymupdf>=1.23.0",
]

[tool.uv]
default-groups = ["dev"]

[dependency-groups]
dev = [
    "pip-licenses>=5.0.0",
    "pre-commit>=4.0.1",
    "pylance>=0.28.0",
    "mypy>=1.15.0",
    "types-openpyxl>=3.1.5.20250516",
    "ruff>=0.11.4",
]
deploy = [
    "twine>=5.1.1",
]

[tool.ruff]
line-length = 120
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = ["uv.lock", "*.yml", ".*", "*.md"]

[tool.hatch.build.targets.wheel]
exclude = ["uv.lock", "*.yml", ".*", "*.md"]
